Add messages and extract dataAnalysis from utils
