+++ manifest.json
@@ -17,6 +17,7 @@
   },
   "builders": {
     "react": "2.x",
-    "pages": "0.x"
+    "pages": "0.x",
+    "messages": "1.x"
   }
 }
