+++ messages/context.json
@@ -1,0 +1,18 @@
+{
+  "console.menu.title": "",
+  "console.submenu.metrics": "",
+  "console.submenu.errors": "",
+  "console.submenu.evidences": "",
+  "console.submenu.publishes": "",
+  "console.admin.metrics.datePicker.startDateTitle": "",
+  "console.admin.metrics.datePicker.endDateTitle": "",
+  "console.admin.components.metrics.controllers": "",
+  "console.admin.tabs.metrics": "",
+  "console.admin.tabs.errors": "",
+  "console.admin.tabs.builds": "",
+  "console.submenu.kpis": "",
+  "console.admin.metrics.button.edit": "",
+  "console.admin.metrics.button.save": "",
+  "console.admin.metrics.button.cancel": "",
+  "console.admin.metrics.button.add": ""
+}
